% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_qc_nfeatures_and_ncounts.R
\name{smart_qc_cutoff}
\alias{smart_qc_cutoff}
\title{自动识别分布形态并计算过滤阈值 (修复版)}
\usage{
smart_qc_cutoff(
  values,
  adjust = 2,
  valley_strictness = 0.99,
  min_peak_ratio = 1e-04,
  assumption = "gaussian",
  show_plot = TRUE,
  title = ""
)
}
\arguments{
\item{values}{数值向量。例如 Seurat 对象的 nFeature_RNA。}

\item{adjust}{平滑度。对于导数计算，平滑非常重要，建议 2.0 或更高。}

\item{valley_strictness}{数值。谷底判定阈值，默认 0.99 (只要谷底比副峰低1\%就算双峰)。}

\item{min_peak_ratio}{数值。副峰高度至少要是主峰高度的多少倍才被考虑？
默认 1e-4 (0.01\%)，用于处理极其悬殊的双峰分布。}

\item{show_plot}{逻辑值。是否绘图。}

\item{title}{字符串。图片标题。}

\item{slope_threshold}{数值。单峰模式下的斜率阈值。}
}
\value{
一个数值 (cutoff)。  并可在Rplot中展示图像(show_plot=TRUE)
}
\description{
这个函数用于根据 nFeature 或 nCount 的分布（双峰或单峰长尾），
自动选择合适的算法（谷底法或拐点法）来计算过滤阈值。
单峰模式下，采用“二阶导数最右侧局部峰值法”，寻找最紧贴主峰的拐角。
}
